working test program of socket communication between stub and controller
